Finished the code rewrite I started friday evening. I wanted to do some cleanup while I was about to write something about it, plus make some preparations for the future.
Now I finalized how Haskell code is supposed to interface with values 
containing holes (variables). I managed to hide the much more 
complicated internal representation (in subsitutions, etc). There is a 
simple invariant on the external representation that must be kept intact 
to ensure correctness/keep sharing.

Polymorphic holes posed a particular complication, because I need some 
dictionary in order to convert an internal hole to its external 
representation, and for polymorphic values the type is not fixed yet, 
thus the dictionary is not yet known. The trick in this case is that as 
long as this type is not known yet, that the value cannot be touched 
either (only propagated around), and then when the conversion is needed, 
this information is known.