Added dictionaries that may be (up to a certain extend) unkown (to be filled in by unifications). These come in handy now I'm writing a type system for Ruler in Ruler, but I guess that another use-case is non-extensible records.
The limitation is that if unification discovers that an unkown needs to be filled in, and there is actually more than one unknown in the dictionary, then it will first unify the unknowns. This bypasses the problem which to choose when you for example unify: union d1 d2 with d3, where d3 is some concrete dictionary and d1 and d2 are yet unknown.

Only the unify function for this particular data structure has to be manually written (because my Template Haskell program can only derive structural equality, and this is insufficient here). The API functions, such as testing for the presence of an item, taking the difference, etc. are expressible in ruler itself (examples/dict.rul).