Some essential debugging. The examples I wanted to put in the tutorial stopped working, since changes the last week.
Spend most of the morning chasing a ghost bug due to a small mistake in the pretty printer.
Fixed a problem that due to a change the unification was not symmetric anymore. Didn't encounter this before.
Haven't discovered the cause of the main problem yet though, but am getting closer...