Changed the ordering of rules. I'm now adding additional edges to the graph, such that a topsort of the graph actually gives the vertices in a total order.
The reason is that the partial order imposed by
attribute dependencies leaves freedom on how to
order unrelated rules. For conventional AGs, this
is not really important. However, I'm allowing
multiple clauses per production. In this case,
I want to exploit this freedom to move pattern
matches to the front, and visits to children to
the back of the visit order (while still
respecting the attr dependencies).

This simple idea turned out to be quite hard to
implement. I even had to resort to two different
graph libraries: the conventional Data.Graph
library to compute stronlgy-connected components
in topologically sorted order, and the FGL
library to dynamically extend graphs.