Finished new dependency generation for rules. The default topsort algorithm did not always do some additional work as I hoped it would. Just value-dependencies proved to be insufficient for conditional definitions of attributes.
In (the ordered variant of) ruler-core you can give sets of
attributes conditional definitions by grouping them in clauses,
and tagging some of the rules as backtrackable. If the pattern
match against the lhs of that rule rails, it then backtracks to
the next clause instead of crashing the program. These rules
need to be evaluated asap, with non-backtrackable rules only
preceeding them if the attributes they provide are needed by a
rhs of a backtrackable-rule. This need is apparent when e.g. you
set an attribute to bottom if some identifier is not in the
environment: an rhs with bottom as definition has no dependencies
at all on any attribute, thus is likely to be ordered before the
evaluation of the condition if only value-dependencies are taken
into account.

In more abstract terms, the problem is:
(1) Given a Direct Acyclic Graph, with the nodes painted in different colors (e.g. Red, Blue, Green), and a total order on the colors (Red < Blue < Green).
(2) I now want to order the nodes preferably on the order on the colors (i.e. all red nodes first, then the blue nodes, etc). 
(3) But, in the resulting order, the nodes must still satisfy the partial order imposed by the DAG
We then impose this total order to the graph by adding additional edges.

Preparation step:
We first ensure that the dependency graph is actually a DAG by
temporarily filtering out edges to break cycles, and adding them
back in the end of the process.
Eliminate those edges E from the graph that make it cyclic.
For an edge e `elem` E, it holds that not removing it from
the graph:
(a) makes it cylic, or
(b) does not increase the connected-ness of nodes
Algorithm:
(1) compute strongly connected components
(2) arbitrarily totally-order the nodes in such component
(3) filter out those edges between nodes in such component
    that are going in a different direction than imposed
    by the order.

The idea I now use is to first order all the elements based on the
total order, and then reluctantly moving elements the least distance
as possible in order to satisfy the value-dependencies.

The naive algorithm is:
(a) create an array A and put each node in it.
(b) sort A based on the total order
(c) fixpoint iterate: find the highest index i, such that node A[i] is in conflict. A node A[i] is in conflict with A[j] (i /= j), if A[i] points to A[j] in the DAG, but i > j. In that case, move A[i] just before A[j] in the graph.
So, the order of (2) is a preference, the order of (3) is a requirement.

This may seem to cause a lot of swapping. However, I think this does not
affect the performance a lot, because the graph is sparse and paths are
short.

Invariant: if in previous steps we moved elements minimally, then
moving A[i] just before A[j] moved A[i] minimally. When A[i] has to
be in front of A[j], and A[j] was moved minimally (by induction), then
the position just before A[j] is a minimal movement.
Invariant: termination is harder to explain. There is probably some
summation based on the rank of a node in the graph, and the distances
between predecessors and successors that I can shown to be decreasing
if the graph is acyclic.

In the actual algorithm:
I implemented the following. I want to insert an element in between
two other elements. However, I do not want to change the indices of
existing elements afterward. To do this, I'm using a sparse array with
double-precision floating points as indices, implemented with a Map.
To insert an item between index a and b, I just calculate a value c,
a < c < b (e.g. c = (a + b) / 2, but there are smarter possibilities),
and use c as index. [when the difference between a and b has become
too small, I reindex the entire array, but this doesn't happen often].
This works like a charm, but seems a bit unconventional.

For the fixpoint-iteration process, I'm keeping a set of nodes still
to examine (initially everyting). A node that is moved, adds all its
successors to this set. By first taking the transitive closure of the
DAG, I can add all (indirect) successors in one go to the set. This
at least causes the inserts to be done directly in one go, and I also
have the impression that it makes the fixpoing computation converge
faster, but I'm not sure about that.

After trying many different algorithms the last week(s),
I hope that this algorithm is sufficiently simple, but also fast
enough. 


Also fixed some scoping bugs. In some corner cases, environments
were not propagated properly, causing some edges in the dependency
graph to be pointing to nothing, and thus the dependency to be
missing. This is very likely to cause rules to be ordered too
early with the above algoritm, resulting in static errors about
identifiers by GHC.